reorgnaize components
