+++ src/components/CourseInfo/Panel.js
@@ -1,0 +1,606 @@
+import {
+  Flex,
+  Text,
+  Stat,
+  StatLabel,
+  StatNumber,
+  HStack,
+  Icon,
+  Box,
+  Button,
+  VStack,
+  StatHelpText,
+  Divider,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverArrow,
+  PopoverCloseButton,
+  PopoverHeader,
+  PopoverBody,
+  Spacer,
+  IconButton,
+  useColorModeValue,
+} from "@chakra-ui/react";
+import React from "react";
+import { PieChart } from "react-minimal-pie-chart";
+import {
+  FaCircle,
+  FaExclamationTriangle,
+  FaQuestionCircle,
+  FaChevronLeft,
+  FaChevronRight,
+  FaInfoCircle,
+} from "react-icons/fa";
+import { IoMdOpen } from "react-icons/io";
+import PTTContentRowContainer from "components/CourseInfo/PTTContentRowContainer";
+import SignUpCard from "components/CourseInfo/SignUpCard";
+import SignUpReportForm from "components/CourseInfo/SignUpReportForm";
+import { useUser } from "@auth0/nextjs-auth0";
+import { useRouter } from "next/router";
+import Image from "next/image";
+import {
+  useCourseEnrollData,
+  useNTURatingData,
+  usePTTReviewData,
+  usePTTExamData,
+  useSyllabusData,
+} from "hooks/useCourseInfo";
+
+function LoadingPanel({ title, ...restProps }) {
+  return (
+    <Flex
+      w="100%"
+      flexDirection="column"
+      justifyContent="center"
+      alignItems="center"
+      {...restProps}
+    >
+      <VStack>
+        <Image
+          alt=""
+          src={"/img/parrot/parrot.gif"}
+          height={32}
+          width={32}
+          layout="fixed"
+        />
+        <Text
+          fontSize="lg"
+          fontWeight="800"
+          color="gray.500"
+          textAlign="center"
+        >
+          {title}
+        </Text>
+      </VStack>
+    </Flex>
+  );
+}
+
+function PanelPlaceholder({
+  title = "暫無資訊",
+  isEmpty = true,
+  ...restProps
+}) {
+  return (
+    <Flex
+      w="100%"
+      flexDirection="column"
+      justifyContent="center"
+      alignItems="center"
+      {...restProps}
+    >
+      <Icon
+        as={isEmpty ? FaQuestionCircle : FaExclamationTriangle}
+        boxSize="32px"
+        color="gray.500"
+      />
+      <Text
+        mt="2"
+        fontSize="lg"
+        fontWeight="800"
+        color="gray.500"
+        textAlign="center"
+      >
+        {title}
+      </Text>
+    </Flex>
+  );
+}
+
+function UnauthenticatedPanel({ ...restProps }) {
+  const router = useRouter();
+  return (
+    <Flex
+      flexDirection={"column"}
+      py={2}
+      my={3}
+      h="100%"
+      w="100%"
+      align="center"
+      justify="center"
+      {...restProps}
+    >
+      <Icon as={FaInfoCircle} boxSize="32px" color="gray.500" />
+      <Text
+        mt="2"
+        fontSize="lg"
+        fontWeight="800"
+        color="gray.500"
+        textAlign="center"
+      >
+        會員專屬功能
+      </Text>
+      <Button
+        rightIcon={<FaChevronRight />}
+        size="md"
+        my={2}
+        py={2}
+        colorScheme="blue"
+        fontSize="md"
+        fontWeight="800"
+        onClick={() => {
+          router.push("/api/auth/login");
+        }}
+      >
+        來去登入
+      </Button>
+    </Flex>
+  );
+}
+
+function PanelWrapper({
+  isLoading,
+  loadingFallback = <LoadingPanel title="載入中..." height="100%" />,
+  isUnauth,
+  unauthFallback = <UnauthenticatedPanel />,
+  children,
+}) {
+  if (isLoading) {
+    return loadingFallback;
+  }
+  if (isUnauth) {
+    return unauthFallback;
+  }
+  return <>{children}</>;
+}
+
+export function SignUpPanel({
+  isLoading,
+  isUnauth,
+  course,
+  SignUpPostData,
+  setSignUpPostData,
+  fetchSignUpPostData,
+  signUpCardIdx,
+  setSignUpCardIdx,
+}) {
+  const textColor = useColorModeValue("text.light", "text.dark");
+  return (
+    <PanelWrapper
+      isLoading={isLoading}
+      isUnauth={isUnauth}
+      loadingFallback={
+        <LoadingPanel title="努力跑加簽大地中..." height="100%" />
+      }
+    >
+      {!SignUpPostData ? (
+        <PanelPlaceholder title="無加簽相關資訊" height="100%" />
+      ) : SignUpPostData.length === 0 ? (
+        <Flex
+          w="100%"
+          h="100%"
+          mt="4"
+          flexDirection="column"
+          justifyContent="center"
+          alignItems={{ base: "start", lg: "center" }}
+        >
+          <PanelPlaceholder title="無加簽相關資訊" h="100%" pt="0" />
+          <HStack w="100%" pr="8" mt="8" justify="end">
+            <SignUpReportForm
+              courseId={course.id}
+              haveSubmitted={SignUpPostData.some((obj) => obj.is_owner)}
+              submitCallback={fetchSignUpPostData}
+            />
+          </HStack>
+        </Flex>
+      ) : (
+        <Flex
+          w="100%"
+          h="100%"
+          mt="4"
+          flexDirection="column"
+          justifyContent="center"
+          alignItems={{ base: "start", lg: "center" }}
+        >
+          <SignUpCard
+            post={SignUpPostData[signUpCardIdx]}
+            SignUpPostData={SignUpPostData}
+            setSignUpPostData={setSignUpPostData}
+            fetchSignUpPostData={fetchSignUpPostData}
+          />
+          <HStack w="100%" pr="8" mt="8">
+            <HStack>
+              <IconButton
+                size="md"
+                variant="ghost"
+                icon={<FaChevronLeft />}
+                onClick={() =>
+                  setSignUpCardIdx(
+                    signUpCardIdx === 0
+                      ? SignUpPostData.length - 1
+                      : signUpCardIdx - 1
+                  )
+                }
+              />
+              <IconButton
+                size="md"
+                variant="ghost"
+                icon={<FaChevronRight />}
+                onClick={() =>
+                  setSignUpCardIdx((signUpCardIdx + 1) % SignUpPostData.length)
+                }
+              />
+              <Text
+                fontSize="sm"
+                fontWeight="800"
+                color={textColor}
+                textAlign="center"
+              >
+                {signUpCardIdx + 1}/{SignUpPostData.length}
+              </Text>
+            </HStack>
+            <Spacer />
+            <SignUpReportForm
+              courseId={course.id}
+              haveSubmitted={SignUpPostData.some((obj) => obj.is_owner)}
+              submitCallback={fetchSignUpPostData}
+            />
+          </HStack>
+        </Flex>
+      )}
+    </PanelWrapper>
+  );
+}
+
+export function EnrollStatusPanel({ courseSerial }) {
+  const { user, isLoading: isAuth0Loading } = useUser();
+  const { data: courseEnrollStatus, isLoading } =
+    useCourseEnrollData(courseSerial);
+  return (
+    <PanelWrapper
+      isLoading={isLoading || isAuth0Loading}
+      isUnauth={!user}
+      loadingFallback={
+        <LoadingPanel title="努力取得資訊中..." height="100%" pt={8} />
+      }
+    >
+      {!courseEnrollStatus ? (
+        <PanelPlaceholder
+          title="無法取得課程即時資訊"
+          isEmpty={false}
+          h="100%"
+          pt="8"
+        />
+      ) : (
+        <Flex
+          w="100%"
+          mt="4"
+          flexDirection="row"
+          justifyContent="center"
+          alignItems={{ base: "start", lg: "center" }}
+          flexWrap="wrap"
+        >
+          <Stat>
+            <StatLabel>選上</StatLabel>
+            <StatNumber>{courseEnrollStatus.enrolled}</StatNumber>
+            <StatHelpText>人</StatHelpText>
+          </Stat>
+          <Stat>
+            <StatLabel>選上外系</StatLabel>
+            <StatNumber>{courseEnrollStatus.enrolled_other}</StatNumber>
+            <StatHelpText>人</StatHelpText>
+          </Stat>
+          <Stat>
+            <StatLabel>登記</StatLabel>
+            <StatNumber>{courseEnrollStatus.registered}</StatNumber>
+            <StatHelpText>人</StatHelpText>
+          </Stat>
+          <Stat>
+            <StatLabel>剩餘</StatLabel>
+            <StatNumber>{courseEnrollStatus.remain}</StatNumber>
+            <StatHelpText>空位</StatHelpText>
+          </Stat>
+        </Flex>
+      )}
+    </PanelWrapper>
+  );
+}
+
+export function NTURatingPanel({ courseId }) {
+  const { user, isLoading: isAuth0Loading } = useUser();
+  const { data: ntuRatingData, isLoading } = useNTURatingData(courseId);
+  return (
+    <PanelWrapper
+      isLoading={isLoading || isAuth0Loading}
+      isUnauth={!user}
+      loadingFallback={
+        <LoadingPanel title="查詢評價中..." height="100%" pt={8} />
+      }
+    >
+      {!ntuRatingData ? (
+        <Flex h="100%" flexDirection="column" alignItems="center">
+          <PanelPlaceholder title="無評價資訊" h="100%" pt="8" />
+          <Button
+            mt="4"
+            colorScheme="blue"
+            variant="outline"
+            size="sm"
+            rightIcon={<IoMdOpen />}
+            onClick={() =>
+              window.open(
+                "https://rating.myntu.me/?referrer=ntucourse_neo",
+                "_blank"
+              )
+            }
+          >
+            前往 NTURating 撰寫評價
+          </Button>
+        </Flex>
+      ) : (
+        <Flex h="100%" flexDirection="column" alignItems="start">
+          <Text fontSize="md" fontWeight="600" color="gray.700">
+            NTURating 上共有 {ntuRatingData.count} 筆評價
+          </Text>
+          <HStack w="100%" justify="space-between" my="2">
+            <Stat>
+              <StatLabel>甜度</StatLabel>
+              <StatNumber>{ntuRatingData.sweety}</StatNumber>
+              <StatHelpText>平均值</StatHelpText>
+            </Stat>
+            <Stat>
+              <StatLabel>涼度</StatLabel>
+              <StatNumber>{ntuRatingData.breeze}</StatNumber>
+              <StatHelpText>平均值</StatHelpText>
+            </Stat>
+            <Stat>
+              <StatLabel>紮實度</StatLabel>
+              <StatNumber>{ntuRatingData.workload}</StatNumber>
+              <StatHelpText>平均值</StatHelpText>
+            </Stat>
+            <Stat>
+              <StatLabel>品質</StatLabel>
+              <StatNumber>{ntuRatingData.quality}</StatNumber>
+              <StatHelpText>平均值</StatHelpText>
+            </Stat>
+          </HStack>
+          <Button
+            colorScheme="blue"
+            variant="outline"
+            size="sm"
+            rightIcon={<IoMdOpen />}
+            onClick={() =>
+              window.open(
+                ntuRatingData.url + "?referrer=ntucourse_neo",
+                "_blank"
+              )
+            }
+          >
+            前往 NTURating 查看該課程評價
+          </Button>
+        </Flex>
+      )}
+    </PanelWrapper>
+  );
+}
+
+export function PTTReviewPanel({ courseId }) {
+  const { user, isLoading: isAuth0Loading } = useUser();
+  const { data: pttReviewData, isLoading } = usePTTReviewData(courseId);
+  return (
+    <PanelWrapper
+      isLoading={isLoading || isAuth0Loading}
+      isUnauth={!user}
+      loadingFallback={
+        <LoadingPanel title="努力爬文中..." height="100%" pt={8} />
+      }
+    >
+      {!pttReviewData || !Array.isArray(pttReviewData) ? (
+        <PanelPlaceholder title="無相關貼文資訊" h="100%" pt="8" />
+      ) : (
+        <PTTContentRowContainer info={pttReviewData} height="150px" />
+      )}
+    </PanelWrapper>
+  );
+}
+
+export function PTTExamPanel({ courseId }) {
+  const { user, isLoading: isAuth0Loading } = useUser();
+  const { data: pttExamData, isLoading } = usePTTExamData(courseId);
+  return (
+    <PanelWrapper
+      isLoading={isLoading || isAuth0Loading}
+      isUnauth={!user}
+      loadingFallback={
+        <LoadingPanel title="努力爬文中..." height="100%" pt={8} />
+      }
+    >
+      {!pttExamData || !Array.isArray(pttExamData) ? (
+        <PanelPlaceholder title="無相關貼文資訊" h="100%" pt="8" />
+      ) : (
+        <PTTContentRowContainer info={pttExamData} height="150px" />
+      )}
+    </PanelWrapper>
+  );
+}
+
+export function SyllabusPanel({ courseId }) {
+  const { data: syllabusData, isLoading } = useSyllabusData(courseId);
+  const headingColor = useColorModeValue("heading.light", "heading.dark");
+  const textColor = useColorModeValue("text.light", "text.dark");
+  const syllabusTitle = {
+    intro: "概述",
+    objective: "目標",
+    requirement: "要求",
+    office_hour: "Office Hour",
+    material: "參考書目",
+    specify: "指定閱讀",
+  };
+  return (
+    <PanelWrapper isLoading={isLoading} isUnauth={null}>
+      {!syllabusData || !syllabusData?.syllabus ? (
+        <PanelPlaceholder title="無課程大綱資訊" h="100%" pt="8" />
+      ) : (
+        <Flex
+          w="100%"
+          my="4"
+          flexDirection="column"
+          justifyContent="start"
+          alignItems="start"
+          wordBreak="break-all"
+          overflow="auto"
+        >
+          {Object.keys(syllabusData?.syllabus ?? {}).map((key) => {
+            const line = syllabusData.syllabus[key].split("\n");
+            const content = line.map((item, index) => {
+              return (
+                <Text
+                  key={syllabusTitle[key] + "content" + index}
+                  mb="0.5"
+                  fontSize="md"
+                  fontWeight="400"
+                  color={textColor}
+                >
+                  {item.trim()}
+                </Text>
+              );
+            });
+
+            return (
+              <React.Fragment key={syllabusTitle[key]}>
+                <Text
+                  flexBasis="20%"
+                  mb="1"
+                  fontSize="lg"
+                  fontWeight="600"
+                  color={headingColor}
+                >
+                  {syllabusTitle[key]}
+                </Text>
+                {syllabusData.syllabus[key] !== "" ? (
+                  content
+                ) : (
+                  <Text
+                    key={syllabusTitle[key] + "content"}
+                    fontSize="md"
+                    fontWeight="400"
+                    color="gray.500"
+                  >
+                    無
+                  </Text>
+                )}
+                <Divider my="4" />
+              </React.Fragment>
+            );
+          })}
+        </Flex>
+      )}
+    </PanelWrapper>
+  );
+}
+
+export function GradePolicyPanel({ courseId }) {
+  const { data: syllabusData, isLoading } = useSyllabusData(courseId);
+  const headingColor = useColorModeValue("heading.light", "heading.dark");
+  const textColor = useColorModeValue("text.light", "text.dark");
+  return (
+    <PanelWrapper
+      isLoading={isLoading}
+      isUnauth={null}
+      loadingFallback={
+        <LoadingPanel title="查看配分中..." height="100%" pt={8} />
+      }
+    >
+      {!syllabusData || !syllabusData.grade ? (
+        <PanelPlaceholder title="無評分相關資訊" h="100%" pt="8" />
+      ) : (
+        <Flex
+          my="4"
+          flexDirection={{ base: "column", lg: "row" }}
+          justifyContent="space-evenly"
+          alignItems="center"
+        >
+          <Box w="200px" h="200px">
+            <PieChart
+              lineWidth={50}
+              label={({ dataEntry }) => dataEntry.value + "%"}
+              labelPosition={75}
+              data={syllabusData.grade}
+              labelStyle={() => ({
+                fill: "white",
+                fontSize: "10px",
+                fontFamily: "sans-serif",
+              })}
+            />
+          </Box>
+          <VStack mt={{ base: 4, lg: 0 }} align="start">
+            {syllabusData.grade.map((item, index) => {
+              const line = item.comment.split("\n");
+              const content = line.map((item, index) => {
+                return (
+                  <Text
+                    key={"SyllabusDataContent" + index}
+                    mb="1"
+                    fontSize="md"
+                    fontWeight="400"
+                    color={textColor}
+                  >
+                    {item.trim()}
+                  </Text>
+                );
+              });
+              return (
+                <Popover key={"SyllabusData" + index}>
+                  <PopoverTrigger>
+                    <HStack justify="start" cursor="pointer">
+                      <Icon as={FaCircle} size="20px" color={item.color} />
+                      <Text fontSize="lg" fontWeight="800" color={item.color}>
+                        {item.value}%
+                      </Text>
+                      <Text fontSize="md" fontWeight="600" color={headingColor}>
+                        {item.title}
+                      </Text>
+                    </HStack>
+                  </PopoverTrigger>
+                  <PopoverContent>
+                    <PopoverArrow />
+                    <PopoverCloseButton />
+                    <PopoverHeader>
+                      <HStack>
+                        <Text fontSize="lg" fontWeight="800" color={item.color}>
+                          {item.value}%
+                        </Text>
+                        <Text fontSize="md" fontWeight="600" color="gray.700">
+                          {item.title}
+                        </Text>
+                      </HStack>
+                    </PopoverHeader>
+                    <PopoverBody>
+                      {item.comment === "" ? (
+                        <Text fontSize="md" fontWeight="400" color="gray.700">
+                          無詳細資訊
+                        </Text>
+                      ) : (
+                        content
+                      )}
+                    </PopoverBody>
+                  </PopoverContent>
+                </Popover>
+              );
+            })}
+          </VStack>
+        </Flex>
+      )}
+    </PanelWrapper>
+  );
+}
